renamed commands to eliminate the -command part of the file
