+++ mud/commands/describe.js
