+++ mud/commands/error.js
