+++ mud/commands/look.js
