+++ mud/commands/rename.js
